Move integrand_logdet_product and integrand_slq_product from slq.py to lanczos.py
