+++ matfree/lanczos.py
@@ -1,7 +1,68 @@
 """All things Lanczos' algorithm."""
 
-from matfree.backend import containers, control_flow, func, linalg, np
+from matfree.backend import containers, control_flow, func, linalg, np, tree_util
 from matfree.backend.typing import Array, Callable, Tuple
+
+
+def integrand_logdet_product(depth, matvec, vecmat, /):
+    r"""Construct the integrand for the log-determinant of a matrix-product.
+
+    Here, "product" refers to $X = A^\top A$.
+    """
+    return integrand_slq_product(np.log, depth, matvec, vecmat)
+
+
+def integrand_slq_product(matfun, depth, matvec, vecmat, /):
+    r"""Construct the integrand for the trace of a function of a matrix-product.
+
+    Instead of the trace of a function of a matrix,
+    compute the trace of a function of the product of matrices.
+    Here, "product" refers to $X = A^\top A$.
+    """
+
+    def quadform(v0, *parameters):
+        v0_flat, v_unflatten = tree_util.ravel_pytree(v0)
+        length = linalg.vector_norm(v0_flat)
+        v0_flat /= length
+
+        def matvec_flat(v_flat):
+            v = v_unflatten(v_flat)
+            Av = matvec(v, *parameters)
+            flat, unflatten = tree_util.ravel_pytree(Av)
+            return flat, tree_util.partial_pytree(unflatten)
+
+        w0_flat, w_unflatten = func.eval_shape(matvec_flat, v0_flat)
+        matrix_shape = (*np.shape(w0_flat), *np.shape(v0_flat))
+
+        def vecmat_flat(w_flat):
+            w = w_unflatten(w_flat)
+            wA = vecmat(w, *parameters)
+            return tree_util.ravel_pytree(wA)[0]
+
+        # Decompose into orthogonal-bidiag-orthogonal
+        algorithm = alg_bidiag_full_reortho(depth, matrix_shape=matrix_shape)
+        output = decompose_fori_loop(
+            v0_flat, lambda v: matvec_flat(v)[0], vecmat_flat, algorithm=algorithm
+        )
+        u, (d, e), vt, _ = output
+
+        # Compute SVD of factorisation
+        B = _bidiagonal_dense(d, e)
+        _, S, Vt = linalg.svd(B, full_matrices=False)
+
+        # Since Q orthogonal (orthonormal) to v0, Q v = Q[0],
+        # and therefore (Q v)^T f(D) (Qv) = Q[0] * f(diag) * Q[0]
+        eigvals, eigvecs = S**2, Vt.T
+        fx_eigvals = func.vmap(matfun)(eigvals)
+        return length**2 * linalg.vecdot(eigvecs[0, :], fx_eigvals * eigvecs[0, :])
+
+    return quadform
+
+
+def _bidiagonal_dense(d, e):
+    diag = linalg.diagonal_matrix(d)
+    offdiag = linalg.diagonal_matrix(e, 1)
+    return diag + offdiag
 
 
 def funm_vector_product_spd(matfun, order, matvec, /):
